do not ask delete if no image

+++ src/common_db/sydFind.cxx
@@ -143,12 +143,14 @@ int main(int argc, char* argv[])
 
   // Delete
   if (args_info.delete_flag) {
+    if (results.size() > 0) {
       std::cout << "Really delete " << results.size() << " element"
                 << (results.size() > 1 ? "s ":" ") << "(y/n) ? ";
       char c;
       std::scanf("%c", &c);
       if (c =='y') db->Delete(results, table_name);
       else { LOG(FATAL) << "Abort."; }
+    }
   }
 
   // This is the end, my friend.
